[svc] Add MessageSender in addition to MessageQueue
It's almost like public phone
